Create multiple pages.
Extract some simple functions to dry up the rendering of multiple
views. Very simple now, must change later.
Extract a layout so that I don’t have to copy boilerplate from page to
page.
Create own code for rendering a view inside a layout.  Might use a
library later but do it on own for now.
Create a second page of content.
Update the links in the nav to go to the new page of content.
TODO make the active part of the nav reflect the current page.